LRCI-4191 Fix TestrayBuild objects to work for Testray2
